+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.2.0-alpha (wip)
 
+* `entity.get_armor`
+
 ## 0.1.0-alpha
 
 * First release!